Print loop and select branch
